269 Moved GetSegmentedRegions to Page from TesseractEngine
Moved GetSegmentedRegions to Page from TesseractEngine as this better
reflects the .net api (all page processing functions should be on Page not
TesseractEngine).
