+++ src/Tesseract/Interop/Constants.cs
@@ -9,5 +9,9 @@ namespace Tesseract.Interop
     {
         public const string LeptonicaDllName = "liblept172";
         public const string TesseractDllName = "libtesseract304";
+        
+        // tesseract uses an int to represent true false values.
+        public const int TRUE = 1;
+        public const int FALSE = 0;
     }
 }
